chore: refactor getColors
