+++ src/components/depth-chart/depth-chart.tsx
@@ -7,12 +7,11 @@ import React, {
   useRef,
 } from "react";
 
-import { Colors } from "../../helpers";
 import { useThrottledResizeObserver } from "../../hooks";
-import { string2hex } from "../../renderer/utils";
 import { ThemeVariant } from "../../types";
 import { Chart } from "./chart";
 import styles from "./depth-chart.module.css";
+import { getColors } from "./helpers";
 
 /**
  * Creates a price formatter
@@ -37,59 +36,6 @@ export const FONT_SIZE = 12;
  * Height of the bottom price axis
  */
 export const AXIS_HEIGHT = FONT_SIZE + 5;
-
-export const GRAY = string2hex(Colors.GRAY);
-
-export interface Colors {
-  buyFill: number;
-  buyStroke: number;
-  sellFill: number;
-  sellStroke: number;
-  backgroundSurface: number;
-  textPrimary: number;
-  textSecondary: number;
-}
-
-function getColors(element: HTMLElement): Colors {
-  const cssStyleDeclaration = getComputedStyle(element);
-
-  return {
-    buyFill: string2hex(
-      cssStyleDeclaration.getPropertyValue("--pennant-color-buy-fill").trim() ||
-        "#16452d"
-    ),
-    buyStroke: string2hex(
-      cssStyleDeclaration
-        .getPropertyValue("--pennant-color-buy-stroke")
-        .trim() || "#26ff8a"
-    ),
-    sellFill: string2hex(
-      cssStyleDeclaration
-        .getPropertyValue("--pennant-color-sell-fill")
-        .trim() || "#800700"
-    ),
-    sellStroke: string2hex(
-      cssStyleDeclaration
-        .getPropertyValue("--pennant-color-sell-stroke")
-        .trim() || "#ff261a"
-    ),
-    textPrimary: string2hex(
-      cssStyleDeclaration
-        .getPropertyValue("--pennant-font-color-base")
-        .trim() || "#ffffff"
-    ),
-    textSecondary: string2hex(
-      cssStyleDeclaration
-        .getPropertyValue("--pennant-font-color-secondary")
-        .trim() || "#fafafa"
-    ),
-    backgroundSurface: string2hex(
-      cssStyleDeclaration
-        .getPropertyValue("--pennant-background-surface-color")
-        .trim() || "#0a0a0a"
-    ),
-  };
-}
 
 export type PriceLevel = {
   price: number;
@@ -163,7 +109,7 @@ export const DepthChart = forwardRef(
       return () => {
         chartRef.current.destroy();
       };
-    // eslint-disable-next-line react-hooks/exhaustive-deps
+      // eslint-disable-next-line react-hooks/exhaustive-deps
     }, []);
 
     // Update chart when dimensions or data change
