+++ src/core/x-axis.ts
@@ -1,4 +1,4 @@
-import { Colors } from "../components/chart";
+import { Colors } from "../components/chart/helpers";
 import { XAxisElement, XAxisTooltipElement } from "../elements";
 import { clearCanvas } from "../helpers";
 import { Interval, ScaleTime } from "../types";
