+++ src/elements/element-y-axis-tooltip.ts
@@ -1,4 +1,4 @@
-import { Colors } from "../components/chart";
+import { Colors } from "../components/chart/helpers";
 import { TICK_LABEL_FONT_SIZE, Y_AXIS_WIDTH } from "../constants";
 import { formatter } from "../helpers";
 import { ScaleLinear, ScaleTime } from "../types";
