+++ src/helpers/helpers-scenegraph.ts
@@ -12,7 +12,7 @@ import {
   isFieldLTPredicate,
   Predicate,
 } from "../vega-lite/predicate";
-import { Colors } from ".";
+import { COLORS } from ".";
 
 export const PADDING_INNER = 0.4;
 
@@ -43,14 +43,14 @@ export function getAreaConfig(
   if (y2 === undefined) {
     return {
       points: data.map((d: any) => [d[x], 0, d[y]]),
-      fill: fill ?? Colors.GRAY,
+      fill: fill ?? COLORS.GRAY,
       line: line,
     };
   }
 
   return {
     points: data.map((d: any) => [d[x], d[y], d[y2]]),
-    fill: fill ?? Colors.GRAY,
+    fill: fill ?? COLORS.GRAY,
     line: line,
   };
 }
@@ -162,8 +162,8 @@ export function getConditionalColor(colorDef: any | undefined) {
 
       color = (datum) =>
         datum[predicate.field] < comparison(datum)
-          ? colorDef.condition.value ?? Colors.GRAY
-          : colorDef.value ?? Colors.GRAY;
+          ? colorDef.condition.value ?? COLORS.GRAY
+          : colorDef.value ?? COLORS.GRAY;
     } else if (isFieldGTPredicate(predicate)) {
       const comparison =
         typeof predicate.gt === "string"
@@ -172,8 +172,8 @@ export function getConditionalColor(colorDef: any | undefined) {
 
       color = (datum) =>
         datum[predicate.field] > comparison(datum)
-          ? colorDef.condition.value ?? Colors.GRAY
-          : colorDef.value ?? Colors.GRAY;
+          ? colorDef.condition.value ?? COLORS.GRAY
+          : colorDef.value ?? COLORS.GRAY;
     } else {
       color = () => null;
     }
